# openapi/relay-backend.yaml
openapi: 3.1.0
info:
  title: Zyvor Relay Native Adapter API
  version: 0.1.0
  description: Minimal backend contract consumed by relay-pubsub.
servers:
  - url: http://relay:9090
paths:
...
  /v1/messages:publish:
    post:
      operationId: publish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, messages]
              properties:
                topic: { type: string }
                messages: { type: array, items: { $ref: '#/components/schemas/NewMessage' } }
      responses: { '200': { description: Published } }
...
components:
  schemas:
...
    NewMessage:
      type: object
      required: [data, attributes, ordering_key]
      properties:
        data: { type: array, items: { type: integer, minimum: 0, maximum: 255 } }
        attributes: { type: object, additionalProperties: { type: string } }
        ordering_key: { type: string }